updated common function location
