+++ src/pages/dashboard/components/PSPRegistrationData.tsx
@@ -9,42 +9,7 @@ import ROUTES from '../../../routes';
 import { isPspBrokerSigned, isPspSigned } from '../../../utils/rbac-utils';
 import { usePermissions } from '../../../hooks/usePermissions';
 import { Party } from '../../../model/Party';
-
-export const commonDetails = (t: any, selectedParty?: Party) => {
-  const companyNameField = () => {
-    switch (selectedParty?.institutionType) {
-      case 'PSP':
-        return selectedParty?.pspData?.legalRegisterName;
-      case 'PA':
-        return selectedParty?.description;
-      case 'PT':
-        return selectedParty?.fiscalCode;
-      default:
-        return undefined;
-    }
-  };
-
-  return (
-    <>
-      <Grid item xs={4}>
-        <Typography variant="body2">{t('dashboardPage.registrationData.name')}</Typography>
-      </Grid>
-      <Grid item xs={8}>
-        <Typography variant="body2" fontWeight={'fontWeightMedium'}>
-          {selectedParty?.description}
-        </Typography>
-      </Grid>
-      <Grid item xs={4}>
-        <Typography variant="body2">{t('dashboardPage.registrationData.companyName')}</Typography>
-      </Grid>
-      <Grid item xs={8}>
-        <Typography variant="body2" fontWeight={'fontWeightMedium'}>
-          {companyNameField()}
-        </Typography>
-      </Grid>
-    </>
-  );
-};
+import { commonDetails } from '../DashboardPage';
 
 const PSPRegistrationData = () => {
   const { t } = useTranslation();
